% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_hc.R
\name{add_custom_hover}
\alias{add_custom_hover}
\title{Transform hover_info list to hc accepter sting and add custom tooltip info}
\usage{
add_custom_hover(hc_map, hover_info)
}
\arguments{
\item{hc_map}{"highchart htmlwidget" object, a base map (produced with draw_hc_map)}

\item{hover_info}{named list, specify columns to show on tooltip hover, names must be column names, elements either empty string or text to show in front of the value}
}
\value{
"highchart htmlwidget"
}
\description{
Transform hover_info list to hc accepter sting and add custom tooltip info
}
