% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_maps.R
\name{add_mappoint}
\alias{add_mappoint}
\title{Adds single mappoint to a higcharter map widglet}
\usage{
add_mappoint(hc_map, points_geojson)
}
\arguments{
\item{hc_map}{"highchart htmlwidget" object, a base map (produced with draw_hc_map)}

\item{points_geojson}{spatial object of class geofeaturecollection (can be obtained with from sf object with prep_geojson)}
}
\value{
highchart htmlwidget
}
\description{
Adds single mappoint to a higcharter map widglet
}
