% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_compatibility.R
\name{check_compatibility}
\alias{check_compatibility}
\title{Check Compatibility of Two Datasets for Comparison}
\usage{
check_compatibility(df1, df2)
}
\arguments{
\item{df1}{The first data frame to be compared.}

\item{df2}{The second data frame to be compared.}
}
\value{
A list containing details about the compatibility of the datasets,
including information on dimension equality and common columns.
}
\description{
Verifies if two datasets are compatible for comparison,
focusing on dimensions and variable names.
}
\examples{
check_compatibility(df1, df2)
}
