% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_by_group.R
\name{compare_by_group}
\alias{compare_by_group}
\title{Compare Datasets by Group}
\usage{
compare_by_group(df1, df2, group_vars)
}
\arguments{
\item{df1}{First dataset for comparison.}

\item{df2}{Second dataset for comparison.}

\item{group_vars}{Variables used for grouping the data.}
}
\value{
A list of comparison results for each group.
}
\description{
Performs dataset comparisons based on specified grouping variables.
Useful for BY-group analysis in datasets.
}
\examples{
compare_by_group(df1, df2, group_vars = c("group_var1", "group_var2"))
}
