% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_observations.R
\name{compare_observations}
\alias{compare_observations}
\title{Compare Observations of Two Datasets}
\usage{
compare_observations(df1, df2)
}
\arguments{
\item{df1}{A data frame representing the first dataset.}

\item{df2}{A data frame representing the second dataset.}
}
\value{
A data frame showing the rows where the datasets differ.
}
\description{
Performs row-wise comparison to identify differences in data values between two datasets.
Highlights rows with differences.
}
\examples{
compare_observations(df1, df2)
}
