% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_variables.R
\name{compare_variables}
\alias{compare_variables}
\title{Compare Variables of Two Datasets}
\usage{
compare_variables(df1, df2)
}
\arguments{
\item{df1}{A data frame representing the first dataset.}

\item{df2}{A data frame representing the second dataset.}
}
\value{
A data frame containing the comparison results of the variables.
}
\description{
Compares the variables (columns) of two datasets in terms of names and data types.
Reports discrepancies in variables between the datasets.
}
\examples{
compare_variables(df1, df2)
}
