% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_in_rad}
\alias{count_in_rad}
\title{Count number of points within radius distance}
\usage{
count_in_rad(radius, data1, lonlat1, data2 = NULL, lonlat2 = NULL)
}
\arguments{
\item{radius}{radius km distance}

\item{data1}{data.frame or data.table containing lon and lat columns}

\item{lonlat1}{character vector of column names in data1 holding coordinates (longittude and lattitude)}

\item{data2}{optional data.frame or data.table containing lon and lat columns, if not given square matrix between points in data1 is computed}

\item{lonlat2}{character vector of column names in data2 holding coordinates (longittude and lattitude)}
}
\value{
data.table
}
\description{
Count number of points within radius distance of given number of kilometers
}
