% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crop_n_cast}
\alias{crop_n_cast}
\title{Crop and cast sf data.frame object}
\usage{
crop_n_cast(sf_obj, bbox, spat_obj_type = NULL)
}
\arguments{
\item{sf_obj}{sf data.frame object}

\item{bbox}{list containing: xmin, ymin, xmax, ymax that denote extent of cropping area}

\item{spat_obj_type}{character, name of type of each geometry in sf data.frame (only needed if cropping causes ambiguity)}
}
\value{
sf object
}
\description{
Crop and cast sf data.frame object
}
