% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_maps.R
\name{draw_hc_full_map}
\alias{draw_hc_full_map}
\title{Draw map with points from base map object and points object}
\usage{
draw_hc_full_map(map_obj, points_obj)
}
\arguments{
\item{map_obj}{spatial object of class geofeaturecollection for base map (can be obtained with from sf object with prep_geojson)}

\item{points_obj}{list of spatial object of class geofeaturecollection, or signle geofeaturecollection object, these denote points to be placed on the map (can be obtained with from sf object with prep_geojson)}
}
\value{
object of class "highchart htmlwidget"
}
\description{
Base map must be of geofeaturecollection class, points_obj can be either a signle object of a list of objects of geofeaturecollection
}
