% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_hc.R
\name{draw_hc_map}
\alias{draw_hc_map}
\title{Draw highchart base map (contours)}
\usage{
draw_hc_map(map_obj)
}
\arguments{
\item{map_obj}{spatial object of class geofeaturecollection (can be obtained with from sf object with prep_geojson)}
}
\value{
object of class "highchart htmlwidget"
}
\description{
Draw highchart base map (contours)
}
