% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_radar.R
\name{draw_radar}
\alias{draw_radar}
\title{Draw plotly radar plot by categories}
\usage{
draw_radar(data, id_var, grouper, mapped_vars, agreg_type, filters = NULL)
}
\arguments{
\item{data}{data.frame or data.table object}

\item{id_var}{character, a column name in data that holds record identifier}

\item{grouper}{character, column name that holds categories that defines groups by which individual records can be compared, if NULL no grouping}

\item{mapped_vars}{character, columns names' which hold measure variables to be displayed on radar}

\item{agreg_type}{type of the aggregation do be displayed, one of available in get_agr function (mean, median, sd); if NULL, no aggregation displayed}

\item{filters}{list objects, it's elements are categorical column names, list values are corresponding values; list for restraining data, so plot show only selected records, not all records in dataset}
}
\value{
plotly object
}
\description{
Draw plotly radar plot for given categories based on given numeric columns
}
