% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_avg}
\alias{get_avg}
\title{Get averages by groups}
\usage{
get_avg(data, grp_col, val_col)
}
\arguments{
\item{data}{data.frame or data.table object}

\item{grp_col}{character vector of column names that specify the groups}

\item{val_col}{character vector of columns names that hold values to be aggregated}
}
\value{
data.table
}
\description{
Aggregates data by given groups to averages
}
