% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_n_closest}
\alias{get_n_closest}
\title{Find records given number of closests records from given one}
\usage{
get_n_closest(n, data1, id1, lonlat1, data2 = NULL, id2 = NULL, lonlat2 = NULL)
}
\arguments{
\item{n}{number of closest locations to be returned}

\item{data1}{data.frame or data.table containing lon and lat columns}

\item{id1}{character of column name holding identifiers in data1}

\item{lonlat1}{character vector of column names in data1 holding coordinates (longittude and lattitude)}

\item{data2}{optional data.frame or data.table containing lon and lat columns, if not given square matrix between points in data1 is computed}

\item{id2}{character of column name holding identifiers in data2}

\item{lonlat2}{character vector of column names in data2 holding coordinates (longittude and lattitude)}
}
\value{
data.table
}
\description{
Find records given number of closests records from given one
}
