% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_n_in_rad}
\alias{get_n_in_rad}
\title{Find records within radius distance}
\usage{
get_n_in_rad(
  radius,
  data1,
  id1,
  lonlat1,
  data2 = NULL,
  id2 = NULL,
  lonlat2 = NULL
)
}
\arguments{
\item{radius}{radius distance in km}

\item{data1}{data.frame or data.table containing lon and lat columns}

\item{id1}{character of column name holding identifiers in data1}

\item{lonlat1}{character vector of column names in data1 holding coordinates (longittude and lattitude)}

\item{data2}{optional data.frame or data.table containing lon and lat columns, if not given square matrix between points in data1 is computed}

\item{id2}{character of column name holding identifiers in data2}

\item{lonlat2}{character vector of column names in data2 holding coordinates (longittude and lattitude)}
}
\value{
data.table
}
\description{
Find records within radius distance
}
