% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rad_dist_matrix}
\alias{get_rad_dist_matrix}
\title{Get matrix of radius distances}
\usage{
get_rad_dist_matrix(data1, lonlat1, data2 = NULL, lonlat2 = NULL)
}
\arguments{
\item{data1}{data.frame or data.table containing lon and lat columns}

\item{lonlat1}{character vector of column names in data1 holding coordinates (longittude and lattitude)}

\item{data2}{optional data.frame or data.table containing lon and lat columns, if not given square matrix between points in data1 is computed}

\item{lonlat2}{character vector of column names in data2 holding coordinates (longittude and lattitude)}
}
\value{
data.table
}
\description{
Compute matrix of radius distances between points in two data.frames/data.table
}
