% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radar.R
\name{get_radar}
\alias{get_radar}
\title{Draw plotly radar plot by categories}
\usage{
get_radar(data, filters, id_var, grouper, mapped_vars, aggregation)
}
\arguments{
\item{data}{data.frame or data.table object}

\item{filters}{list objects where column names are list elements and categories are corresponding values, data gets filtered using a list}

\item{id_var}{character, a column name in data that holds record identifier}

\item{grouper}{character, column name that holds categories for grouping data, if NULL no grouping}

\item{mapped_vars}{character, columns names' which hold measure variables to be displayed on radar}

\item{aggregation}{type of the aggregation do be displayed, one of avaliables in get_agr function (mean, median, sd); if NULL, no aggregation displayed}
}
\value{
plotly object
}
\description{
Draw plotly radar plot for given categories based on given numeric columns
}
