% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huBvizoRkit-package.R
\docType{package}
\name{huBvizoRkit-package}
\alias{huBvizoRkit}
\alias{huBvizoRkit-package}
\title{huBvizoRkit: What the Package Does (One Line, Title Case)}
\description{
Package contaningin handy functions for data preparation and visualizations, with intetion to use in shiny app modules
}
\author{
\strong{Maintainer}: Hubert Smieciuch \email{hubert.cooperates@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
