% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype.R
\name{hype}
\alias{hype}
\title{Hyperparameter optimization}
\description{
Hyperparameter optimization

Hyperparameter optimization
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{hype$new()}}
\item \href{#method-add_data}{\code{hype$add_data()}}
\item \href{#method-add_X}{\code{hype$add_X()}}
\item \href{#method-add_LHS}{\code{hype$add_LHS()}}
\item \href{#method-add_EI}{\code{hype$add_EI()}}
\item \href{#method-run_all}{\code{hype$run_all()}}
\item \href{#method-run_EI_for_time}{\code{hype$run_EI_for_time()}}
\item \href{#method-plotorder}{\code{hype$plotorder()}}
\item \href{#method-plotX}{\code{hype$plotX()}}
\item \href{#method-print}{\code{hype$print()}}
\item \href{#method-clone}{\code{hype$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hype$new(eval_func, ..., X0 = NULL, n_lhs, extract_output_func)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_data"></a>}}
\if{latex}{\out{\hypertarget{method-add_data}{}}}
\subsection{Method \code{add_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hype$add_data(X, Y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_X"></a>}}
\if{latex}{\out{\hypertarget{method-add_X}{}}}
\subsection{Method \code{add_X()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hype$add_X(X)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_LHS"></a>}}
\if{latex}{\out{\hypertarget{method-add_LHS}{}}}
\subsection{Method \code{add_LHS()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hype$add_LHS(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_EI"></a>}}
\if{latex}{\out{\hypertarget{method-add_EI}{}}}
\subsection{Method \code{add_EI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hype$add_EI(n, covtype = "matern5_2", nugget.estim = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all"></a>}}
\if{latex}{\out{\hypertarget{method-run_all}{}}}
\subsection{Method \code{run_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hype$run_all(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_EI_for_time"></a>}}
\if{latex}{\out{\hypertarget{method-run_EI_for_time}{}}}
\subsection{Method \code{run_EI_for_time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hype$run_EI_for_time(
  sec,
  batch_size,
  covtype = "matern5_2",
  nugget.estim = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotorder"></a>}}
\if{latex}{\out{\hypertarget{method-plotorder}{}}}
\subsection{Method \code{plotorder()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hype$plotorder()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plotX"></a>}}
\if{latex}{\out{\hypertarget{method-plotX}{}}}
\subsection{Method \code{plotX()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hype$plotX()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hype$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hype$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
