% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbc.R
\name{mbc}
\alias{mbc}
\title{Microbenchmark compare}
\usage{
mbc(..., times = 5, input, inputi, post, target)
}
\arguments{
\item{...}{Functions to run}

\item{times}{Number of times to run}

\item{input}{Object to be passed as input to each function}

\item{inputi}{Function to be called with the replicate number then passed to each function.}

\item{post}{Function to post-process results.}

\item{target}{Values the functions are expected to (approximately) return.}
}
\value{
Data frame of comparison results
}
\description{
Microbenchmark compare
}
\examples{
m1 <- mbc(function(x) {Sys.sleep(rexp(1, 30));mean(x)},
  function(x) {Sys.sleep(rexp(1, 5));median(x)}, input=runif(100))
mbc(function(x) {Sys.sleep(rexp(1, 30));mean(x)},
  function(x) {Sys.sleep(rexp(1, 5));median(x)}, input=runif(100),
  post=function(x){c(x+1, 12)})
mbc(function(x) {Sys.sleep(rexp(1, 30));mean(x)+runif(1)},
  function(x) {Sys.sleep(rexp(1, 5));median(x)+runif(1)}, input=runif(100),
  post=function(x){c(x+1, 12)}, times=3)
m1 <- mbc(function() {x <- runif(100);Sys.sleep(rexp(1, 30));mean(x)},
  function() {x <- runif(100);Sys.sleep(rexp(1, 5));median(x)})
}
