% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype_par_discretenum.R
\name{par_discretenum}
\alias{par_discretenum}
\title{Parameter with uniform distribution for hyperparameter optimization}
\usage{
par_discretenum(name, values)
}
\arguments{
\item{name}{Name of the parameter, must match the input to `eval_func`.}

\item{values}{Values, discrete numeric}
}
\description{
Parameter with uniform distribution for hyperparameter optimization
}
\examples{
p1 <- par_discretenum('x1', 0:2)
class(p1)
print(p1)
}
