% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype.R
\name{par_hype}
\alias{par_hype}
\title{Parameter for hyperparameter optimization}
\description{
Parameter for hyperparameter optimization

Parameter for hyperparameter optimization
}
\examples{
p1 <- par_hype$new()
class(p1)
print(p1)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{par_hype$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_hype$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
