% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype_par_integer.R
\name{par_integer}
\alias{par_integer}
\title{Parameter with uniform distribution over integer range
for hyperparameter optimization}
\usage{
par_integer(name, lower, upper)
}
\arguments{
\item{name}{Name of the parameter, must match the input to `eval_func`.}

\item{lower}{Lower bound of the parameter}

\item{upper}{Upper bound of the parameter}
}
\description{
Parameter with uniform distribution over integer range
for hyperparameter optimization
}
\examples{
p1 <- par_integer('x1', 3, 8)
class(p1)
print(p1)
table(p1$generate(runif(1000)))
}
