% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype_par_log.R
\name{par_log10}
\alias{par_log10}
\title{Hyperparameter on log10 scale}
\description{
Hyperparameter on log10 scale

Hyperparameter on log10 scale
}
\section{Super class}{
\code{\link[comparer:par_hype]{comparer::par_hype}} -> \code{par_log10}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fromraw}{\code{par_log10$fromraw()}}
\item \href{#method-toraw}{\code{par_log10$toraw()}}
\item \href{#method-new}{\code{par_log10$new()}}
\item \href{#method-clone}{\code{par_log10$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromraw"></a>}}
\if{latex}{\out{\hypertarget{method-fromraw}{}}}
\subsection{Method \code{fromraw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_log10$fromraw(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toraw"></a>}}
\if{latex}{\out{\hypertarget{method-toraw}{}}}
\subsection{Method \code{toraw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_log10$toraw(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a hyperparameter with uniform distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_log10$new(name, lower, upper)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the parameter, must match the input to `eval_func`.}

\item{\code{lower}}{Lower bound of the parameter}

\item{\code{upper}}{Upper bound of the parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_log10$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
