% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype_par_ordered.R
\name{par_ordered}
\alias{par_ordered}
\title{Hyperparameter of discrete (factor) variable}
\usage{
par_ordered(name, values)
}
\arguments{
\item{name}{Name of the parameter, must match the input to `eval_func`.}

\item{values}{Vector of values}
}
\description{
Hyperparameter of discrete (factor) variable
}
\examples{
p1 <- par_ordered('x1', c('a', 'b', 'c'))
class(p1)
print(p1)
}
