% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbc_clean.R
\name{plot.mbc}
\alias{plot.mbc}
\title{Plot mbc class}
\usage{
\method{plot}{mbc}(x, ...)
}
\arguments{
\item{x}{Object of class mbc}

\item{...}{Additional parameters}
}
\value{
None
}
\description{
Plot mbc class
}
\examples{
m1 <- mbc(mn= {Sys.sleep(rexp(1, 30));mean(x)},
          med={Sys.sleep(rexp(1, 5));median(x)},
          input=runif(100))
plot(m1)
}
