% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_geojson}
\alias{prep_geojson}
\title{Transform sf of data.frame to geojson of type FeatureCollection}
\usage{
prep_geojson(data, lon, lat, crs, bbox = NULL, spat_obj_type = NULL)
}
\arguments{
\item{data}{data.frame, data.table or sf frame containing data}

\item{lon}{character, longitude column name}

\item{lat}{character, latitude column name}

\item{crs}{EPSG CRS code of Coordinate Reference Systems}

\item{bbox}{list containing: xmin, ymin, xmax, ymax that denote extent of cropping area}

\item{spat_obj_type}{haracter, name of type of each geometry in sf data.frame (only needed if cropping causes ambiguity)}
}
\value{
FeatureCollection object
}
\description{
Transform sf of data.frame to geojson of type FeatureCollection used for drawing highcharter map.
}
