% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prep_sf}
\alias{prep_sf}
\title{Transform data.frame or data.table to sf object}
\usage{
prep_sf(data, lon, lat, crs, bbox = NULL, spat_obj_type = NULL)
}
\arguments{
\item{data}{data.frame or data.table to be transformed to sf object}

\item{lon}{character, longitude column name}

\item{lat}{character, latitude column name}

\item{crs}{EPSG CRS code of Coordinate Reference Systems}

\item{bbox}{list containing: xmin, ymin, xmax, ymax that denote extent of cropping area}

\item{spat_obj_type}{haracter, name of type of each geometry in sf data.frame (only needed if cropping causes ambiguity)}
}
\value{
sf object
}
\description{
Transform data.frame or data.table to sf object
}
