% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_hc.R
\name{raise_if_not_geojson}
\alias{raise_if_not_geojson}
\title{Check if object is geofeaturecollection, otherwise raise error}
\usage{
raise_if_not_geojson(object)
}
\arguments{
\item{object}{object the check is performed against}
}
\description{
Check if object is geofeaturecollection, otherwise raise error
}
