% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_data}
\alias{scale_data}
\title{Scale data to be within given range}
\usage{
scale_data(data, num_col, min = 0, max = 10)
}
\arguments{
\item{data}{data.frame or data.table object}

\item{num_col}{character vector of columns names that hold values to be scaled}

\item{min}{min of a new range}

\item{max}{max of a new range}
}
\value{
data.table
}
\description{
Agregates data by given groups to averages
}
