% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_county_mainland_sf.R
\docType{data}
\name{usa_county_mainland_sf}
\alias{usa_county_mainland_sf}
\title{US countries sf data.frame class object restricted by bbox}
\format{
sf data.frame with 3108 rows and 7 variables:
\describe{
\item{REGIONCE}{}
\item{AFFGEOID}{}
\item{GEOID}{}
\item{NAME}{}
\item{LSAD}{}
\item{ALAND}{}
\item{AWATER}{}
\item{geometry}{}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
usa_county_mainland_sf
}
\description{
US countries sf data.frame class object restricted to: min lon-125.200195 degrees, min lat 23.523700 degrees, max lon -66.533203 degrees, max lat 49.439557 degrees, without Alaska, Hawaii and Puerto Rico.
}
\keyword{datasets}
