% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_county_sf.R
\docType{data}
\name{usa_county_sf}
\alias{usa_county_sf}
\title{US counties sf data.frame class object}
\format{
sf data.frame with 3220 rows and 7 variables:
\describe{
\item{REGIONCE}{}
\item{AFFGEOID}{}
\item{GEOID}{}
\item{NAME}{}
\item{LSAD}{}
\item{ALAND}{}
\item{AWATER}{}
\item{geometry}{}
}
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html}
}
\usage{
usa_county_sf
}
\description{
US counties sf data.frame class object
}
\keyword{datasets}
