% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_data.R
\docType{data}
\name{usa_data}
\alias{usa_data}
\title{Dataset created based on datasets::state.x77}
\format{
A data.table with 50 rows and 14 variables:
\describe{
\item{state}{state name}
\item{region}{state region}
\item{lon}{centroid longitude}
\item{lat}{centroid latitude}
\item{rand_status}{}
\item{rand_class}{}
\item{Population}{}
\item{Income}{}
\item{Illiteracy}{}
\item{Life Exp}{}
\item{Murder}{}
\item{HS Grad}{}
\item{Frost}{}
\item{Area}{}
}
}
\source{
\url{datasets::state.x77}
}
\usage{
usa_data
}
\description{
Dataset created based on datasets::state.x77
}
\keyword{datasets}
