% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usa_sf.R
\docType{data}
\name{usa_sf}
\alias{usa_sf}
\title{US states sf data.frame class object}
\format{
sf data.frame with 52 rows and 7 variables:
\describe{
\item{geoid}{}
\item{state_name}{}
\item{state_abbr}{}
\item{jurisdiction_type}{}
\item{aland}{}
\item{awater}{}
\item{geometry}{}
}
}
\source{
\url{USAboundaries::states_contemporary_lores}
}
\usage{
usa_sf
}
\description{
US states sf data.frame class object
}
\keyword{datasets}
