% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_maps.R
\name{verify_points_object}
\alias{verify_points_object}
\title{Verify if object only consists of geofeaturecollection, and return it if so}
\usage{
verify_points_object(points_obj)
}
\arguments{
\item{points_obj}{object the check is performed against}
}
\value{
geofeaturecollection
}
\description{
Verify if object only consists of geofeaturecollection, and return it if so
}
