% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_maps.R
\name{add_hc_mappoins}
\alias{add_hc_mappoins}
\title{Add points series(es) to highchart htmlwidget base contour map}
\usage{
add_hc_mappoins(hc_map, points_obj)
}
\arguments{
\item{hc_map}{"highchart htmlwidget" object, a base map (produced with draw_hc_map)}

\item{points_obj}{list of spatial object of class geofeaturecollection, or signle geofeaturecollection object, these denote points to be placed on the map (can be obtained with from sf object with prep_geojson)}
}
\value{
object of class "highchart htmlwidget"
}
\description{
Add points series(es) to highchart htmlwidget base contour map
}
