% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_hc.R
\name{draw_hc_full_map}
\alias{draw_hc_full_map}
\title{Draw map with points from base map object and points object}
\usage{
draw_hc_full_map(map_obj, points_obj, points_names = NULL)
}
\arguments{
\item{map_obj}{spatial object of class geofeaturecollection for base map (can be obtained with from sf object with prep_geojson)}

\item{points_obj}{list of spatial object of class geofeaturecollection, or single geofeaturecollection object, these denote points to be placed on the map (can be obtained with from sf object with prep_geojson)}

\item{points_names}{character vector containing names for each series to be drawn}
}
\value{
object of class "highchart htmlwidget"
}
\description{
Base map must be of geofeaturecollection class, points_obj can be either a single object of a list of objects of geofeaturecollection
}
