% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype_par_discrete.R
\name{par_discrete}
\alias{par_discrete}
\title{Hyperparameter of discrete (factor) variable}
\description{
Hyperparameter of discrete (factor) variable

Hyperparameter of discrete (factor) variable
}
\examples{
p1 <- par_discrete$new('x1', c('a', 'b', 'c'))
class(p1)
print(p1)
}
\section{Super class}{
\code{\link[comparer:par_hype]{comparer::par_hype}} -> \code{par_discrete}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the parameter, must match the input to `eval_func`.}

\item{\code{values}}{Vector of values}

\item{\code{ggtrans}}{Transformation for ggplot, see ggplot2::scale_x_continuous()}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-fromraw}{\code{par_discrete$fromraw()}}
\item \href{#method-toraw}{\code{par_discrete$toraw()}}
\item \href{#method-fromint}{\code{par_discrete$fromint()}}
\item \href{#method-toint}{\code{par_discrete$toint()}}
\item \href{#method-generate}{\code{par_discrete$generate()}}
\item \href{#method-getseq}{\code{par_discrete$getseq()}}
\item \href{#method-new}{\code{par_discrete$new()}}
\item \href{#method-clone}{\code{par_discrete$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromraw"></a>}}
\if{latex}{\out{\hypertarget{method-fromraw}{}}}
\subsection{Method \code{fromraw()}}{
Function to convert from raw scale to transformed scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_discrete$fromraw(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Value of raw scale}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toraw"></a>}}
\if{latex}{\out{\hypertarget{method-toraw}{}}}
\subsection{Method \code{toraw()}}{
Function to convert from transformed scale to raw scale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_discrete$toraw(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Value of transformed scale}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromint"></a>}}
\if{latex}{\out{\hypertarget{method-fromint}{}}}
\subsection{Method \code{fromint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_discrete$fromint(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toint"></a>}}
\if{latex}{\out{\hypertarget{method-toint}{}}}
\subsection{Method \code{toint()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_discrete$toint(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate"></a>}}
\if{latex}{\out{\hypertarget{method-generate}{}}}
\subsection{Method \code{generate()}}{
Generate values in the raw space based on quantiles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_discrete$generate(q)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q}}{In [0,1].}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getseq"></a>}}
\if{latex}{\out{\hypertarget{method-getseq}{}}}
\subsection{Method \code{getseq()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_discrete$getseq(n)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a hyperparameter with uniform distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_discrete$new(name, values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the parameter, must match the input to `eval_func`.}

\item{\code{values}}{The values the variable can take on.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{par_discrete$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
