% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype_par_log.R
\name{par_log10}
\alias{par_log10}
\title{Hyperparameter on log10 scale}
\usage{
par_log10(name, lower, upper)
}
\arguments{
\item{name}{Name of the parameter, must match the input to `eval_func`.}

\item{lower}{Lower bound of the parameter}

\item{upper}{Upper bound of the parameter}
}
\description{
Hyperparameter on log10 scale
}
\examples{
p1 <- par_log10('x1', 1e-4, 1e4)
class(p1)
print(p1)
}
