% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype_par_unif.R
\name{par_unif}
\alias{par_unif}
\title{Uniform parameter}
\usage{
par_unif(name, lower, upper)
}
\arguments{
\item{name}{Name of the parameter, must match the input to `eval_func`.}

\item{lower}{Lower bound of the parameter}

\item{upper}{Upper bound of the parameter}
}
\value{
Returns an R6 class generated by R6_par_unif.
}
\description{
Parameter with uniform distribution for hyperparameter optimization
}
\examples{
p1 <- par_unif('x1', 1, 10)
class(p1)
print(p1)
}
