% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbc_clean.R
\name{print.mbc}
\alias{print.mbc}
\title{Print mbc class}
\usage{
\method{print}{mbc}(x, ...)
}
\arguments{
\item{x}{Object of class mbc}

\item{...}{Additional parameters}
}
\value{
None
}
\description{
Print mbc class
}
\examples{
m1 <- mbc({Sys.sleep(rexp(1, 30));mean(x)},
          {Sys.sleep(rexp(1, 5));median(x)},
          input=runif(100))
print(m1)
}
