% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_hc.R
\name{verify_points_object}
\alias{verify_points_object}
\title{Verify if object only consists of geofeaturecollection, and return it if so}
\usage{
verify_points_object(points_obj, points_names)
}
\arguments{
\item{points_obj}{object the check is performed against}

\item{points_names}{character vector containing names for each series to be drawn}
}
\value{
geofeaturecollection
}
\description{
Verify if object only consists of geofeaturecollection, and return it if so
}
